2004-11-20	Thomas Doerfler <Thomas.Doerfler@imd-systems.de>
        PR 703/filesystem
	* ide/ide.c: Extend the "modes_available" and "mode_active" parameters
	to uint16_t for UDMA support.

+++ c/src/lib/libbsp/i386/pc386/ide/ide.c
@@ -288,8 +288,8 @@ rtems_status_code pc386_ide_config_io_speed
 +---------------------------------------------------------------------------+
 | Input Parameters:                                                         |
 \*-------------------------------------------------------------------------*/
- int       minor,                   /* controller minor number       */
- uint8_t   modes_avail              /* optional argument             */
+ int        minor,                   /* controller minor number       */
+ uint16_t   modes_avail              /* optional argument             */
  )
 /*-------------------------------------------------------------------------*\
 | Return Value:                                                             |
